pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mizan-flow"
version = "0.1.0"
description = "Mizan Flow - Product Lifecycle Management Platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt<4.1",
    "python-multipart>=0.0.18",
    "httpx>=0.28.0",
    "redis>=5.2.0",
    "boto3>=1.36.0",
    "openai>=1.58.0",
    "anthropic>=0.40.0",
    "aiofiles>=24.1.0",
    "google-auth>=2.0.0",
    "requests>=2.31.0",
    "resend>=2.0.0",
    "google-cloud-storage>=2.14.0",
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
    "fpdf2>=2.8.0",
    "arq>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.setuptools.packages.find]
include = ["apps*", "packages*"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
